Address review: inline upload helpers, use push_to_hub
- Move generate_pairs and build_pair_dataset into the upload script
  (they were only used at upload time, not at task evaluation time)
- Drop mteb/tasks/pair_classification/eng/_video_pair_helpers.py
- Replace manual parquet shard upload with datasets.Dataset.push_to_hub
- Drop the rst :mod: cross-reference
